move prisma folder into src
