+++ package.json
@@ -62,7 +62,8 @@
     "vite": "^5.1.4"
   },
   "prisma": {
-    "seed": "tsx -r dotenv/config prisma/seed.ts"
+    "schema": "src/prisma/schema.prisma",
+    "seed": "tsx -r dotenv/config src/prisma/seed.ts"
   },
   "engines": {
     "node": ">=18.0.0"
