+++ src/access/app-db/game.ts
@@ -1,5 +1,5 @@
 import { Game, Prisma } from '@prisma/client';
-import { appDB } from '../../../prisma';
+import { appDB } from '../../prisma';
 
 export async function createGame(gameData: Prisma.GameCreateArgs['data']): Promise<Game> {
   return appDB.game.create({
