+++ src/access/app-db/official-game.ts
@@ -1,5 +1,5 @@
 import { Prisma } from '@prisma/client';
-import { appDB } from '../../../prisma';
+import { appDB } from '../../prisma';
 
 export async function createOfficialGames(officialGameData: Prisma.OfficialGameCreateManyArgs['data']): Promise<void> {
   await appDB.officialGame.createMany({
