+++ src/access/app-db/official.ts
@@ -1,5 +1,5 @@
 import { Official, Prisma } from '@prisma/client';
-import { appDB } from '../../../prisma';
+import { appDB } from '../../prisma';
 
 export async function getOfficialByExternalId(externalId: string): Promise<Official | null> {
   return appDB.official.findFirst({
