+++ src/access/app-db/player-game.ts
@@ -1,5 +1,5 @@
 import { PlayerGame, Prisma } from '@prisma/client';
-import { appDB } from '../../../prisma';
+import { appDB } from '../../prisma';
 
 export async function createPlayerGame(playerGameData: Prisma.PlayerGameCreateArgs['data']): Promise<PlayerGame> {
   return appDB.playerGame.create({
