+++ src/access/app-db/team.ts
@@ -1,5 +1,5 @@
 import { Team } from '@prisma/client';
-import { appDB } from '../../../prisma';
+import { appDB } from '../../prisma';
 
 export async function getTeamByExternalId(externalId: string): Promise<Team> {
   return appDB.team.findFirstOrThrow({
